Consolidate osprofiler options
Right now every project has its own copy of the profiler options,
This situation will make it very hard to keep up when we want to
change things in osprofiler. So we should consolidate the code
just like all other oslo libraries in osprofiler itself. However
one twist here is that we should avoid needing oslo.config as a
hard requirement for project that do not want to use it, so we
should add this as an extra in setup.cfg and adjust the testenv
as well.

Change-Id: I4343136be501551bb2c87d4b66d839c4e63a67fc