stop making a copy of options discovered by config generator
In order for the config generator's new hook for updating default values
to work, we have to return the same Opt objects from list_opts() that
our set_defaults() function is going to modify. There's no real need to
be making copies of the objects anyway, so just stop doing it.

Change-Id: I9c6149c354adc1827b773aa07bf5874981ee8b8f
Signed-off-by: Doug Hellmann <doug@doughellmann.com>