run py34 tests before py27 to work around testr bug
testr uses anydbm to create a database, but python 3 dropped support for
the database format that is the default in python 2. The default for
python 3 is available under python 2, so if we run the python 3 tests
before the python 2 tests everything works fine. Change the order in the
default environment list so that if someone checks out the repo and runs
'tox' all of the tests work.

Change-Id: I44b23add5acc297ca00d2d3c7747f0cfff85cc2f
Signed-off-by: Doug Hellmann <doug@doughellmann.com>