Add functional test for notifier backend
We have several notifier backends now. It is important to verify
that they are not broken when changes are made. Adding a test for
notifier backend verification. We can use this test to verify the
backends against each osprofiler patch.

Note: the support for messaging/ceilometer backend will be added
in a subsequent patch.

Change-Id: Icb0f453de30dbfa1a7a990e73391cf17a246e9bb