Highlight last trace for OSprofiler html output
Sometime, exception or error occurs, we want to find problems quickly.
However, if an output of OSprofiler is long, it will very hard to find
where is the last trace that get stuck.

This visualization can help us a lot in finding problems in services.

Example: https://tovin07.github.io/glance-image-list-highlight.html

Change-Id: Iaabb7850db4dfeec8ce38e0ef5264db8122355ef