Make collector configurable in DevStack plugin
In DevStack OSProfiler is by default configured to use messaging
driver with Ceilometer backend. User can change the driver, but
still needs to install collector/storage manually. This patch
introduces configuration option `OSPROFILER_COLLECTOR` which is
used to specify which collector/storage to install into DevStack.

Currently 2 values are supported:
 * `redis` to install Redis server and use Redis driver
 * <empty> to keep the default behavior and use messaging driver

To test the patch on DevStack, the following lines are needed in local.conf:
  enable_plugin osprofiler https://git.openstack.org/openstack/osprofiler refs/changes/06/527406/5
  OSPROFILER_BRANCH=refs/changes/06/527406/5
  OSPROFILER_COLLECTOR=redis

Change-Id: Ia9763605db95b3f35c8b0e51211f96ee0dd3a82d