Add minimum version and fix dulwich issue
Main changes:
- Add minimum version for requirements [1]
- Fix dulwich issue that broke py35 gate [2]

Trivial changes:
- Re-format multi-line deps in tox.ini
- Verify and change constraints of pymongo, prettytable
  to match global requirements

---
requirements without minimum version definitaions may cause CI failure

Merge https://review.openstack.org/#/c/574367/ made requirements job
check for minimum version definitions, requirements without minimum version
may cause CI failure.

This patch sync the requirements with lower-constraints

---
Start using upper-constraints in tox venvs

Without this we are exposed to any breakages in dependencies, which
is currently blocking the repo because of a broken dulwich release.

Note that this required the removal of the explicit .[oslo_config]
requirement.  I'm not sure why that was there since it will be
pulled in by oslo.concurrency anyway and it broke the venv creation
when constraints were added.  Since I'm not sure why it was split
out as a separate thing I'm leaving the setup.cfg target but
removing it from tox.ini.

We also need to pin dulwich in lower-constraints, even though it's a
transitive dependency.  Otherwise the lower-constraints job installs
the broken latest version.

Transitive dependency that we need to pin due to
https://bugs.launchpad.net/tripleo/+bug/1778004
---

[1] https://review.openstack.org/#/c/575691
[2] https://review.openstack.org/#/c/577462

Co-Authored-By: Kevin_Zheng <zhengzhenyu@huawei.com>
Co-Authored-By: Ben Nemec <bnemec@redhat.com>

Change-Id: Ibfe5f460139849ba39b369321360f6f5a0e9e2a0
Closes-Bug: #1777083
Closes-Bug: #1778004