Make profiler.clean() public method
Horizon dashboard has a profiler plugin which calls profiler.init
method at
 openstack_dashboard/contrib/developer/profiler/middleware.py
line 64. This creates a profiler and picks a new base_id. The only
way to remove this profiler and stop using this base_id is to call
profiler._clean() method. So this patch changes _clean() to clean()
so it becomes a public method.

Change-Id: Idec7ee240bc7f508aeebcba374a9673652b1cc72
Related-Bug: #1777486