Add sqlalchemy collector
Beside the already available collectors, add a sqlalchemy based
collector. This is useful if you don't want to maintain another DB
solution and just use the (usually) already available database.
The driver currently implements the notify() and get_report() methods
so it is possible to store trace points and to get a single trace.

Change-Id: If91b35d4b97862c0ecf6677f4c6b95a09d411195