Use upper constraints
Sphinx 2.0.0 dropped python 2.7 and 3.4 support. Because osprofiler
uses global upper-constraints.txt only from Rocky,
job openstack-tox-py27 tries to install Sphinx 2.0.1 and fails [1] with:

 Downloading http://mirror.iad.rax.openstack.org/wheel/ubuntu-16.04-x86_64/sphinx/Sphinx-2.0.1-py2.py3-none-any.whl (3.2MB)
 ERROR: Sphinx requires Python '>=3.5' but the running Python is 2.7.12

Note: .[oslo.config] could be removed as the packages will be installed
perfectly and constrained without this and it breaks installation if
upper-constraints.txt is used anyway.

[1] http://logs.openstack.org/periodic-stable/opendev.org/openstack/osprofiler/stable/queens/openstack-tox-py27/3d469d3/job-output.txt.gz#_2019-05-10_06_10_56_192617

Change-Id: Ife6bcaca347d914cb1fc7e2fe71665ea5c93ca80