Bring env OSPROFILER_CONNECTION_STRING into effect
The --connection-string is marked required, and parse_args()
assumes the flag invalid if it's not supplied as command line
argument, even if env OSPROFILER_CONNECTION_STRING is setted.

We should not rely on the required=True of argparse module,
but check the required args on our own to take env arguments
into consideration.

Change-Id: I50f5f369ec41ddc82f9b66f5910b7805940fb387