List up extra requirements for drivers
Most of the drivers require extra requirements. This adds these to
the extras field in setup.cfg so that users can more easily install
extra requirements.

Also removes the lower-constraints.txt file because we already removed
tests of l-c by [1].

Note that the lower versions of SQLAlchemy and oslo.messaging are
picked up from requirements in the other repos, and further lower
versions might be allowed.

[1] bcca5bd4a8c3fed96dfcf01bf24233fa509e610b

Change-Id: If4492943199c75c294e4eb5a3600cf5d2bdad0b8